Reduce fs ops for dev bundle patch apply
